Make nsync talk to bbs instead of receptor
[#102265598]

+++ recipebuilder/docker_execution_metadata.go
@@ -11,7 +11,7 @@ type DockerExecutionMetadata struct {
 }
 
 type Port struct {
-	Port     uint16
+	Port     uint32
 	Protocol string
 }
